Refactor: Updated the Code to Handle Multiple Directories inside Each Other with Correct Output Forwarding
The needed files to be minified are now found using *find* in bash. Output is *grepped* to avoid minified files. Then we get the output path, name, file extension of the output file, and run the npm script on the files.

The new approach fixes the bug represented in issue #1. In addition, there is a commit referenced in that issue to a failed workflow for that commit

This refactor fixes issue #1, makes the code cleaner, and more modular